+++ Device/Simulator/CommandExtractor.cs
@@ -9,14 +9,14 @@
             _dirPath = dirPath;
         }
 
-        public byte[] ExtractCommand(byte code)
+        public byte[] ExtractCommand(short code)
         {
             var file = FindFileCommand(code);
 
             return file == null ? Array.Empty<byte>() : Convert.FromHexString(File.ReadAllText(file));
         }
 
-        private string FindFileCommand(byte code)
+        private string FindFileCommand(short code)
         {
             return Directory.GetFiles(_dirPath, $"{code:X2}.txt").FirstOrDefault();
         }
